Logins examples in docs.

+++ views.py
@@ -1,7 +1,10 @@
-from flask import render_template, current_app, abort, request, redirect, url_for
+from flask import render_template, current_app, abort, request, redirect, url_for, flash
 from datetime import datetime
 from movie import Movie
-from forms import MovieEditForm
+from forms import MovieEditForm, LoginForm
+from flask_login import login_user, logout_user, current_user, login_required
+from user import get_user
+from passlib.hash import pbkdf2_sha256 as hasher
 
 
 def home_page():
@@ -16,9 +19,12 @@ def movies_page():
         movies = db.get_movies()
         return render_template("movies.html", movies=sorted(movies))
     else:
+        if not current_user.is_admin:
+            abort(401)
         form_movie_keys = request.form.getlist("movie_keys")
         for form_movie_key in form_movie_keys:
             db.delete_movie(int(form_movie_key))
+        flash("%(num)d movies deleted." % {"num": len(form_movie_keys)})
         return redirect(url_for("movies_page"))
 
 
@@ -30,63 +36,10 @@ def movie_page(movie_key):
     return render_template("movie.html", movie=movie)
 
 
+@login_required
 def movie_add_page():
-    if request.method == "GET":
-        values = {"title": "", "year": ""}
-        return render_template(
-            "movie_edit.html",
-            min_year=1887,
-            max_year=datetime.now().year,
-            values=values,
-        )
-    else:
-        valid = validate_movie_form(request.form)
-        if not valid:
-            return render_template(
-                "movie_edit.html",
-                min_year=1887,
-                max_year=datetime.now().year,
-                values=request.form,
-            )
-        title = request.form.data["title"]
-        year = request.form.data["year"]
-        movie = Movie(title, year=year)
-        db = current_app.config["db"]
-        movie_key = db.add_movie(movie)
-        return redirect(url_for("movie_page", movie_key=movie_key))
-
-
-def movie_edit_page(movie_key):
-    if request.method == "GET":
-        db = current_app.config["db"]
-        movie = db.get_movie(movie_key)
-        if movie is None:
-            abort(404)
-        values = {"title": movie.title, "year": movie.year}
-        return render_template(
-            "movie_edit.html",
-            min_year=1887,
-            max_year=datetime.now().year,
-            values=values,
-        )
-    else:
-        valid = validate_movie_form(request.form)
-        if not valid:
-            return render_template(
-                "movie_edit.html",
-                min_year=1887,
-                max_year=datetime.now().year,
-                values=request.form,
-            )
-        title = request.form.data["title"]
-        year = request.form.data["year"]
-        movie = Movie(title, year=year)
-        db = current_app.config["db"]
-        db.update_movie(movie_key, movie)
-        return redirect(url_for("movie_page", movie_key=movie_key))
-
-
-def movie_add_page():
+    if not current_user.is_admin:
+        abort(401)
     form = MovieEditForm()
     if form.validate_on_submit():
         title = form.data["title"]
@@ -94,10 +47,12 @@ def movie_add_page():
         movie = Movie(title, year=year)
         db = current_app.config["db"]
         movie_key = db.add_movie(movie)
+        flash("Movie added.")
         return redirect(url_for("movie_page", movie_key=movie_key))
     return render_template("movie_edit.html", form=form)
 
 
+@login_required
 def movie_edit_page(movie_key):
     db = current_app.config["db"]
     movie = db.get_movie(movie_key)
@@ -107,7 +62,30 @@ def movie_edit_page(movie_key):
         year = form.data["year"]
         movie = Movie(title, year=year)
         db.update_movie(movie_key, movie)
+        flash("Movie data updated.")
         return redirect(url_for("movie_page", movie_key=movie_key))
     form.title.data = movie.title
     form.year.data = movie.year if movie.year else ""
     return render_template("movie_edit.html", form=form)
+
+
+def login_page():
+    form = LoginForm()
+    if form.validate_on_submit():
+        username = form.data["username"]
+        user = get_user(username)
+        if user is not None:
+            password = form.data["password"]
+            if hasher.verify(password, user.password):
+                login_user(user)
+                flash("You have logged in.")
+                next_page = request.args.get("next", url_for("home_page"))
+                return redirect(next_page)
+        flash("Invalid credentials.")
+    return render_template("login.html", form=form)
+
+
+def logout_page():
+    logout_user()
+    flash("You have logged out.")
+    return redirect(url_for("home_page"))
